// src/types/partials/Identifier.ts
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TUri } from '../simpleTypes/Uri';
import { TPeriod } from '../partials/Period';
import { TReference } from '../partials/Reference';

export type TIdentifier = {
    id?: String;
    extension?: TExtension[];
    use?: String;
    type?: TCodeableConcept;
    system?: TUri;
    value?: String;
    period?: TPeriod;
    assigner?: TReference;
};