// src/types/partials/Coding.ts
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TUri } from '../simpleTypes/Uri';

export type TCoding = {
    id?: String;
    extension?: TExtension[];
    system?: TUri;
    version?: String;
    code?: String;
    display?: String;
    userSelected?: Boolean;
};